ARK-003 반환값 분리
