+++ core/core-api/src/main/kotlin/io/ark/springboot/core/api/controller/v1/FileController.kt
@@ -1,6 +1,7 @@
 package io.ark.springboot.core.api.controller.v1
 
-import io.ark.springboot.core.domain.file.FileDto
+import io.ark.springboot.core.api.controller.v1.response.FileDownloadUrlResponse
+import io.ark.springboot.core.api.controller.v1.response.FileResponse
 import io.ark.springboot.core.domain.file.FileService
 import io.ark.springboot.core.support.response.ApiResponse
 import io.ark.springboot.storage.db.core.file.UploadStatus
@@ -49,26 +50,3 @@ class FileController(
     }
 }
 
-data class FileResponse(
-    val id: Long,
-    val originalName: String,
-    val size: Long,
-    val mimeType: String,
-    val status: UploadStatus,
-    val url: String?,
-) {
-    companion object {
-        fun from(dto: FileDto, url: String?) = FileResponse(
-            id = dto.id,
-            originalName = dto.originalName,
-            size = dto.size,
-            mimeType = dto.mimeType,
-            status = dto.status,
-            url = url,
-        )
-    }
-}
-
-data class FileDownloadUrlResponse(
-    val url: String,
-)
