+++ core/core-api/src/main/kotlin/io/ark/springboot/core/api/controller/v1/response/FileDownloadUrlResponse.kt
@@ -1,0 +1,5 @@
+package io.ark.springboot.core.api.controller.v1.response
+
+data class FileDownloadUrlResponse(
+    val url: String,
+)
